Soyoon

+++ src/main/java/com/example/demo/repository/MemberRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberRepository extends JpaRepository<User, Long> {
+public interface MemberRepository extends JpaRepository<User, Integer> {
 }
